- use posix_shm module for data segment in SharedMemoryManager - use numeric ID instead of string ID for SHM buffer identification - this feature speeds up segment lookups and reduces wire overhead - remove unnecessary fields from SharedMemoryManager - fix clippy warnings - added comments

+++ commons/zenoh-shm/src/watchdog/confirmator.rs
@@ -138,7 +138,7 @@ impl WatchdogConfirmator {
             std::collections::btree_map::Entry::Occupied(occupied) => occupied.get().clone(),
         };
 
-        let index = (descriptor.index_and_bitpos >> 6) as u32;
+        let index = descriptor.index_and_bitpos >> 6;
         let bitpos = descriptor.index_and_bitpos & 0x3f;
 
         let atomic = unsafe { segment.array.elem(index) };
